сейчас добавлю роли

+++ lit/src/main/java/com/litsite/lit/controller/ChapterController.java
@@ -14,27 +14,27 @@ import java.util.List;
 public class ChapterController {
     private final ChapterService chapterService;
     @GetMapping("/{bookId}/chapters")
-    public List<ChapterDto> getChapters(@PathVariable int bookId) {
+    public List<ChapterDto> getChapters(@PathVariable long bookId) {
         return chapterService.getChapters(bookId);
     }
 
     @PostMapping("/{bookId}/chapters")
-    public ChapterDto createChapter(@PathVariable int bookId, @RequestBody CreateChapterDto chapterDto) {
+    public ChapterDto createChapter(@PathVariable long bookId, @RequestBody CreateChapterDto chapterDto) {
         return chapterService.createChapter(bookId, chapterDto);
     }
 
     @GetMapping("/chapters/{chapterId}")
-    public ChapterDto getChapter(@PathVariable int chapterId) {
+    public ChapterDto getChapter(@PathVariable long chapterId) {
         return chapterService.getChapter(chapterId);
     }
 
     @PutMapping("/chapters/{chapterId}")
-    public ChapterDto updateChapter(@PathVariable int chapterId, @RequestBody ChapterDto chapterDto) {
+    public ChapterDto updateChapter(@PathVariable long chapterId, @RequestBody ChapterDto chapterDto) {
         return chapterService.updateChapter(chapterId, chapterDto);
     }
 
     @DeleteMapping("/chapters/{chapterId}")
-    public void deleteChapter(@PathVariable int chapterId) {
+    public void deleteChapter(@PathVariable long chapterId) {
         chapterService.deleteChapter(chapterId);
     }
 }
